Create a point indexer for prometheus based on series id
The label set for each timeseries is used to determine
an index for each point in order to more closely simulate
how prometheus' QueueManager distributes the samples
to different queues for sending to the remote endpoints.

+++ pkg/targets/akumuli/scan.go
@@ -36,10 +36,10 @@ type pointIndexer struct {
 	nchan uint
 }
 
-func (i *pointIndexer) GetIndex(p *data.LoadedPoint) int {
+func (i *pointIndexer) GetIndex(p *data.LoadedPoint) uint {
 	hdr := p.Data.([]byte)
 	id := binary.LittleEndian.Uint32(hdr[0:4])
-	return int(id % uint32(i.nchan))
+	return uint(id) % i.nchan
 }
 
 type batch struct {
